ecypher.js: Ecypher can build, encypher, decypher
`ecypher.js` is a tool for building repeating letter cyphers from a
given alphabet where the first letter of the alphabet is represented by
one repeating letter, the second letter by two repeating letters, the
third by three repeating letters, etc.

For instance, a repeating-letter cypher for using the letter `e` and the
English alphabet would be a=e, b=ee, c=eee, d=eeee, etc.

The module has functions for building cyphers with a given repeating letter and
alphabet `buildCypher`, encyphering with a given message and cypher
`encypher`, and decyphering with a given message and cypher `decypher`.
The module can be tested with the `test` function.